# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestorProfileRead": {
        "additionalProperties": false,
        "description": "The stored plan, plus any warnings it implies.",
        "properties": {
          "drawdown_tolerance_pct": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Drawdown Tolerance Pct",
            "type": "string"
          },
          "emergency_fund_in_place": {
            "title": "Emergency Fund In Place",
            "type": "boolean"
          },
          "goals_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals Note"
          },
          "investable_capital": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Investable Capital",
            "type": "string"
          },
          "investing_borrowed_money": {
            "title": "Investing Borrowed Money",
            "type": "boolean"
          },
          "max_position_pct": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Max Position Pct",
            "type": "string"
          },
          "max_sector_pct": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Max Sector Pct",
            "type": "string"
          },
          "review_interval_days": {
            "title": "Review Interval Days",
            "type": "integer"
          },
          "risk_tolerance": {
            "$ref": "#/components/schemas/RiskTolerance"
          },
          "time_horizon": {
            "$ref": "#/components/schemas/TimeHorizon"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "time_horizon",
          "risk_tolerance",
          "drawdown_tolerance_pct",
          "investable_capital",
          "max_position_pct",
          "max_sector_pct",
          "emergency_fund_in_place",
          "investing_borrowed_money",
          "review_interval_days"
        ],
        "title": "InvestorProfileRead",
        "type": "object"
      },
      "InvestorProfileUpsert": {
        "additionalProperties": false,
        "description": "Create or replace the investor's plan.\n\nModelled as a full replacement (PUT) rather than a patch: this document is\nshort, and the guide's instruction is to write the whole thing down before\nbuying anything. A partial update would let a user quietly raise their own\nrisk limit without re-reading the rest.",
        "properties": {
          "drawdown_tolerance_pct": {
            "anyOf": [
              {
                "maximum": 100.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "30",
            "description": "The drop you could hold through without selling in a panic.",
            "title": "Drawdown Tolerance Pct"
          },
          "emergency_fund_in_place": {
            "default": false,
            "title": "Emergency Fund In Place",
            "type": "boolean"
          },
          "goals_note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Goals Note"
          },
          "investable_capital": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "0",
            "description": "Money you can afford to have tied up or lose. Not rent or emergency funds.",
            "title": "Investable Capital"
          },
          "investing_borrowed_money": {
            "default": false,
            "title": "Investing Borrowed Money",
            "type": "boolean"
          },
          "max_position_pct": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "15",
            "description": "Largest share of the portfolio any single holding may take.",
            "title": "Max Position Pct"
          },
          "max_sector_pct": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 100.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "35",
            "description": "Largest share of the portfolio any single sector may take.",
            "title": "Max Sector Pct"
          },
          "review_interval_days": {
            "default": 90,
            "description": "How often to revisit the reason you own each holding.",
            "maximum": 730.0,
            "minimum": 7.0,
            "title": "Review Interval Days",
            "type": "integer"
          },
          "risk_tolerance": {
            "$ref": "#/components/schemas/RiskTolerance"
          },
          "time_horizon": {
            "$ref": "#/components/schemas/TimeHorizon"
          }
        },
        "required": [
          "time_horizon",
          "risk_tolerance"
        ],
        "title": "InvestorProfileUpsert",
        "type": "object"
      },
...
      "TradeCreate": {
        "additionalProperties": false,
        "description": "Record an executed trade.",
        "properties": {
          "executed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Defaults to now. Accepts a past date for back-filled trades.",
            "title": "Executed At"
          },
          "fees": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "default": "0",
            "description": "Brokerage, CDC and taxes.",
            "title": "Fees"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The plan this trade acts on, if there is one.",
            "title": "Plan Id"
          },
          "price": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "description": "Price per share, excluding fees.",
            "title": "Price"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Quantity"
          },
          "side": {
            "$ref": "#/components/schemas/TradeSide"
          },
          "symbol": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol",
          "side",
          "quantity",
          "price"
        ],
        "title": "TradeCreate",
        "type": "object"
      },
...
      "TradeRead": {
        "additionalProperties": false,
        "description": "A recorded trade.",
        "properties": {
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "executed_at": {
            "format": "date-time",
            "title": "Executed At",
            "type": "string"
          },
          "fees": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Fees",
            "type": "string"
          },
          "gross_value": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Gross Value",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "net_cash_flow": {
            "description": "Signed cash effect: negative for a buy, positive for a sell.",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Net Cash Flow",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "price": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price",
            "type": "string"
          },
          "quantity": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Quantity",
            "type": "string"
          },
          "side": {
            "$ref": "#/components/schemas/TradeSide"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "id",
          "symbol",
          "company_name",
          "side",
          "quantity",
          "price",
          "fees",
          "gross_value",
          "net_cash_flow",
          "executed_at"
        ],
        "title": "TradeRead",
        "type": "object"
      },
      "TradeSide": {
        "description": "Direction of an executed trade.",
        "enum": [
          "buy",
          "sell"
        ],
        "title": "TradeSide",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WatchlistItemCreate": {
        "additionalProperties": false,
        "description": "Add a company to the watchlist.\n\n``research_note`` is required with a real minimum length. The guide's first\nlisted mistake is chasing hype, and the cheapest available guard is making\nthe user articulate why they are watching something before the app will\ntrack it.",
        "properties": {
          "research_note": {
            "description": "Why this company is worth watching, in your own words.",
            "maxLength": 4000,
            "minLength": 10,
            "title": "Research Note",
            "type": "string"
          },
          "symbol": {
            "maxLength": 16,
            "minLength": 1,
            "title": "Symbol",
            "type": "string"
          },
          "target_entry_price": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Price at or below which you would consider buying.",
            "title": "Target Entry Price"
          }
        },
        "required": [
          "symbol",
          "research_note"
        ],
        "title": "WatchlistItemCreate",
        "type": "object"
      },
      "WatchlistItemRead": {
        "additionalProperties": false,
        "description": "A watched company with a current price snapshot.",
        "properties": {
          "company_id": {
            "title": "Company Id",
            "type": "integer"
          },
          "company_name": {
            "title": "Company Name",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "distance_to_target_pct": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance To Target Pct"
          },
          "entry_price_reached": {
            "default": false,
            "title": "Entry Price Reached",
            "type": "boolean"
          },
          "has_trade_plan": {
            "default": false,
            "title": "Has Trade Plan",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_close": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Close"
          },
          "last_close_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Close Date"
          },
          "research_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Note"
          },
          "sector": {
            "$ref": "#/components/schemas/Sector"
          },
          "sector_label": {
            "title": "Sector Label",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "target_entry_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Entry Price"
          }
        },
        "required": [
          "id",
          "company_id",
          "symbol",
          "company_name",
          "sector",
          "sector_label",
          "created_at"
        ],
        "title": "WatchlistItemRead",
        "type": "object"
      },
      "WatchlistItemUpdate": {
        "additionalProperties": false,
        "description": "Revise a watchlist entry.",
        "properties": {
          "research_note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Note"
          },
          "target_entry_price": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Entry Price"
          }
        },
        "title": "WatchlistItemUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nDecision-support API for evaluating Pakistan Stock Exchange (PSX) equities.\n\n**What this API does:** scores a company against a fixed, published fundamentals\nchecklist; reports technical readings as timing context; records the pre-buy\nchecklist, profit target and stop-loss you commit to *before* buying; tracks your\nportfolio against the diversification limits you set for yourself; and tells you\nwhen one of your own rules has been crossed.\n\n**What it does not do:** predict prices, rate stocks, or tell you what to buy or\nsell. There is no endpoint that returns a recommendation, and that is a deliberate\nproduct constraint rather than a missing feature.\n\nCheck `GET /api/v1/meta` for `provider.is_synthetic`. When it is `true`, every\nfinancial figure and price served by this API is **generated demonstration data**\nfor real PSX ticker symbols - useful for exercising the application, useless for\nreal decisions.\n",
    "title": "PSX Invest API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/portfolio/trades": {
      "get": {
        "description": "The trade ledger, newest first.",
        "operationId": "list_trades_api_v1_portfolio_trades_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TradeRead"
                  },
                  "title": "Response List Trades Api V1 Portfolio Trades Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List recorded trades",
        "tags": [
          "portfolio"
        ]
      },
      "post": {
        "description": "Append a buy or sell to the ledger.\n\nA buy against a plan that is READY marks that plan EXECUTED: the commitment has\nbeen acted on, and its profit target and stop-loss now govern real money. If no\n``plan_id`` is supplied, a READY plan for the same company is linked\nautomatically.",
        "operationId": "record_trade_api_v1_portfolio_trades_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradeCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradeRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown symbol."
          },
          "422": {
            "description": "Selling more shares than are held, or an invalid plan reference."
          }
        },
        "summary": "Record a trade",
        "tags": [
          "portfolio"
        ]
      }
    },
    "/api/v1/profile": {
      "get": {
        "description": "Your recorded goals, risk limits and money-hygiene declarations.\n\nReturns ``204 No Content`` when no profile exists, rather than a 404 or an\nobject of defaults. A 404 would suggest something is broken, and returning\ndefaults would let the user believe they had already written down goals they\nnever chose - the first thing the guide asks them to do.",
        "operationId": "get_profile_api_v1_profile_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "204": {
            "description": "No profile has been written yet - the client should prompt for one."
          }
        },
        "summary": "Get the investor profile",
        "tags": [
          "investor profile"
        ]
      },
      "put": {
        "description": "Write down your time horizon, risk tolerance and position limits.\n\nA full replacement rather than a patch: the guide's instruction is to decide all\nof this together, before looking at any company. The response includes derived\nwarnings - for example if no emergency fund is recorded, or if the stated risk\ntolerance and drawdown tolerance contradict each other.",
        "operationId": "upsert_profile_api_v1_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestorProfileUpsert"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestorProfileRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "description": "Limits are incoherent, e.g. a position cap above the sector cap."
          }
        },
        "summary": "Create or replace the investor profile",
        "tags": [
          "investor profile"
        ]
      }
    },
    "/api/v1/watchlist": {
      "get": {
        "description": "Companies you are researching, with a current price snapshot on each.\n\n``entry_price_reached`` and ``distance_to_target_pct`` are computed against the\nlatest stored close, so the client does not have to replicate the arithmetic.",
        "operationId": "list_watchlist_api_v1_watchlist_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WatchlistItemRead"
                  },
                  "title": "Response List Watchlist Api V1 Watchlist Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List watched companies",
        "tags": [
          "watchlist"
        ]
      },
      "post": {
        "description": "Add a company to the watchlist.\n\nA research note is required. The guide's first listed mistake is chasing hype,\nand having to write down *why* before the app will track something is the\ncheapest available guard against it.",
        "operationId": "add_watchlist_item_api_v1_watchlist_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Unknown symbol."
          },
          "409": {
            "description": "Already on the watchlist."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Watch a company",
        "tags": [
          "watchlist"
        ]
      }
    },
    "/api/v1/watchlist/{item_id}": {
      "delete": {
        "description": "Remove a company from the watchlist.",
        "operationId": "remove_watchlist_item_api_v1_watchlist__item_id__delete",
        "parameters": [
          {
            "description": "Watchlist entry id.",
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "description": "Watchlist entry id.",
              "minimum": 1,
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Entry not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop watching a company",
        "tags": [
          "watchlist"
        ]
      },
      "patch": {
        "description": "Revise the research note or the entry price you are waiting for.",
        "operationId": "update_watchlist_item_api_v1_watchlist__item_id__patch",
        "parameters": [
          {
            "description": "Watchlist entry id.",
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "description": "Watchlist entry id.",
              "minimum": 1,
              "title": "Item Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Entry not found."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update a watchlist entry",
        "tags": [
          "watchlist"
        ]
      }
    }
  }
}